:truck: move: Google 광고 관련 로직 advertisement 도메인으로 이동
