+++ src/main/java/com/whereyouad/WhereYouAd/domains/advertisement/presentation/docs/GoogleAdDocs.java
@@ -1,6 +1,6 @@
-package com.whereyouad.WhereYouAd.domains.platform.presentation.docs;
+package com.whereyouad.WhereYouAd.domains.advertisement.presentation.docs;
 
-import com.whereyouad.WhereYouAd.domains.platform.application.dto.response.GoogleAdResponse;
+import com.whereyouad.WhereYouAd.domains.advertisement.application.dto.response.GoogleAdResponse;
 import com.whereyouad.WhereYouAd.global.response.DataResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
